moved js file into utils
